Squashed commit of the following:

commit fac1f35b403b72b0a37c011e930d7a24d2ebe60c
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Tue Feb 20 14:56:15 2018 +0100

    Fix Unit Tests for NativeFileSizeStatistics

commit e6c1519d55a630adbb35ad6229ca8b55cc06047d
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Tue Feb 20 14:54:19 2018 +0100

    [REL-201888] Fix Integration points error Job status failed

+++ Source/kCura.IntegrationPoints.Data.Tests/Statistics/NativeFileSizeStatisticsTests.cs
@@ -60,7 +60,7 @@ namespace kCura.IntegrationPoints.Data.Tests.Statistics
 			var queryResult = MockQueryResult(artifactIds);
 			_relativityObjectManager.Query(Arg.Any<QueryRequest>()).Returns(queryResult);
 
-			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<int>(
+			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<long>(
 					_SQL_TEXT,
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@ArtifactIds" && x.TypeName == "IDs"),
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@FileType" && (FileType)x.Value == FileType.Native))
@@ -88,7 +88,7 @@ namespace kCura.IntegrationPoints.Data.Tests.Statistics
 			var queryResult = MockQueryResult(artifactIds);
 			_relativityObjectManager.Query(Arg.Any<QueryRequest>()).Returns(queryResult);
 
-			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<int>(
+			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<long>(
 					_SQL_TEXT,
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@ArtifactIds" && x.TypeName == "IDs"),
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@FileType" && (FileType)x.Value == FileType.Native))
@@ -116,7 +116,7 @@ namespace kCura.IntegrationPoints.Data.Tests.Statistics
 			var queryResult = MockQueryResult(artifactIds, ProductionConsts.DocumentFieldGuid);
 			_relativityObjectManager.Query(Arg.Any<QueryRequest>()).Returns(queryResult);
 
-			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<int>(
+			_helper.GetDBContext(_WORKSPACE_ID).ExecuteSqlStatementAsScalar<long>(
 					_SQL_TEXT,
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@ArtifactIds" && x.TypeName == "IDs"),
 					Arg.Is<SqlParameter>(x => x.ParameterName == "@FileType" && (FileType)x.Value == FileType.Native))
