+++ Source/kCura.IntegrationPoints.Data/Statistics/Implementations/NativeFileSizeStatistics.cs
@@ -97,7 +97,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 
 		private int GetArtifactId(RelativityObject relativityObject) => relativityObject.ArtifactID;
 
-		private int GetTotalFileSize(IEnumerable<int> artifactIds, int workspaceArtifactId)
+		private long GetTotalFileSize(IEnumerable<int> artifactIds, int workspaceArtifactId)
 		{
 			const string sqlText = "SELECT COALESCE(SUM([Size]),0) FROM [File] WHERE [Type] = @FileType AND [DocumentArtifactID] IN (SELECT * FROM @ArtifactIds)";
 
@@ -113,7 +113,7 @@ namespace kCura.IntegrationPoints.Data.Statistics.Implementations
 			};
 
 			IDBContext dbContext = _helper.GetDBContext(workspaceArtifactId);
-			return dbContext.ExecuteSqlStatementAsScalar<int>(sqlText, artifactIdsParameter, fileTypeParameter);
+			return dbContext.ExecuteSqlStatementAsScalar<long>(sqlText, artifactIdsParameter, fileTypeParameter);
 		}
 
 	}
